feat: Split TC002 into TC002 + TC003 and drop old TC003
The TC003 error seems like was a mistake to introduce in retrospect.
There are plenty of reasons for maintaining separate type-checking
blocks in longer files. So for v2 I've wanted to drop this error.

At the same time, the TC002 error has previously spanned built-in
and third-party imports. With the new classification, we now need
an additional error code for built-in (and future) imports. TC003
seems like a good error code to occupy since the other two are listed
as TC001 and TC002.