feat: Add `TC006` to check `typing.cast()`
This adds a check to ensure that the type argument passed to
`typing.cast()` is a string literal. Unquoted values have no use at
runtime as `typing.cast()` is effectively a no-op and have negative
impacts on performance for complex generic types.

In addition, when unquoted, the types are required to be available at
runtime, even when using `from __future__ import annotations` to opt
into PEP 563. When quoted, those types can then be moved into
`TYPE_CHECKING` blocks.

Closes #114.