feat: Support use of TYPE_CHECKING with simple boolean logic
For various reasons, modules may be conditioned on multiple criteria that determine whether they should be imported.
For example, if we have an module that should be imported during testing or type checking, we may do something like:
```python
if TYPE_CHECKING or environment == 'test':
    import foobar
```

Currently, the linter will flag the `foobar` import as needing to be moved into an `if TYPE_CHECKING` block in this scenario.
This gives us two options, each of which is less than ideal.
We could import the module twice, i.e.
```python
if TYPE_CHECKING:
    import foobar
elif environment == 'test':
    import foobar
```
or use a `# noqa` comment and lose the safety of the check when the code changes.

This commit adds some logic to handle these types of simple conditionals and determine if they are equivalent to `if TYPE_CHECKING:` for the purposes of this linter.
In our simple example, we should be able to determine that `if TYPE_CHECKING or environment == 'test'` will always result in the module being imported if `TYPE_CHECKING` is True.
Similarly, we should be able to determine that `if TYPE_CHECKING and environment == 'test'` is _not_ sufficient, since we have no guarantees about the value of `environment`.